requirements.txt
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
discord-webhook==0.9.0
django-braces==1.14.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
drf-nested-routers==0.91
idna==2.10
importlib-metadata==1.6.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
